requirements.txt
Django>=5.1,<6.0
mysqlclient>=2.2
Pillow>=10.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-redis>=5.4
kafka-python>=2.0
celery>=5.4
redis>=5.0
django-celery-beat>=2.6
flower>=2.0
